Removing a few implicit widening conversions

+++ src/lib/comper.cc
@@ -56,7 +56,7 @@ class comper_internal {
         // marker sequence.
         constexpr static size_t const NumTermBits = 1;
         // Number of bits for end-of-file marker.
-        constexpr static size_t const TerminatorWeight = NumTermBits + 2 * 8;
+        constexpr static size_t const TerminatorWeight = NumTermBits + size_t(2) * 8;
         // Flag that tells the compressor that new descriptor fields is needed
         // when a new bit is needed and all bits in the previous one have been
         // used up.
@@ -132,7 +132,7 @@ public:
             } else {
                 // Dictionary match.
                 // Distance and length of match.
-                size_t const distance = (0x100 - src.getbyte()) * 2;
+                size_t const distance = (size_t(0x100) - src.getbyte()) * 2;
                 size_t const length   = src.getbyte();
                 if (length == 0) {
                     break;
